Optimized obj loading and seperated the quad and obj models into seperate files
