+++ eduRend.vcxproj
@@ -158,6 +158,8 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="lib\stb_image.h" />
+    <ClInclude Include="src\OBJModel.h" />
+    <ClInclude Include="src\QuadModel.h" />
     <ClInclude Include="src\RenderPass.h" />
     <ClInclude Include="src\Camera.h" />
     <ClInclude Include="src\Drawcall.h" />
@@ -176,10 +178,11 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\Camera.cpp" />
-    <ClCompile Include="src\Model.cpp" />
     <ClCompile Include="src\InputHandler.cpp" />
     <ClCompile Include="src\Main.cpp" />
     <ClCompile Include="src\OBJLoader.cpp" />
+    <ClCompile Include="src\OBJModel.cpp" />
+    <ClCompile Include="src\QuadModel.cpp" />
     <ClCompile Include="src\Scene.cpp" />
     <ClCompile Include="src\shader.c" />
     <ClCompile Include="src\Texture.cpp" />
