+++ src/Model.h
@@ -62,56 +62,7 @@ public:
 	}
 };
 
-/**
- * @brief Model reprsenting a Quad.
-*/
-class QuadModel : public Model
-{
-	unsigned nbr_indices = 0;
 
-public:
-	QuadModel(
-		ID3D11Device* dx3ddevice,
-		ID3D11DeviceContext* dx3ddevice_context);
 
-	virtual void Render() const;
-
-	~QuadModel() { }
-};
-
-/**
- * @brief Model representing a 3D object.
- * @see OBJLoader
-*/
-class OBJModel : public Model
-{
-	// index ranges, representing drawcalls, within an index array
-	struct IndexRange
-	{
-		unsigned int start;
-		unsigned int size;
-		unsigned ofs;
-		int mtl_index;
-	};
-
-	std::vector<IndexRange> index_ranges;
-	std::vector<Material> materials;
-
-	void append_materials(const std::vector<Material>& mtl_vec)
-	{
-		materials.insert(materials.end(), mtl_vec.begin(), mtl_vec.end());
-	}
-
-public:
-
-	OBJModel(
-		const std::string& objfile,
-		ID3D11Device* dxdevice,
-		ID3D11DeviceContext* dxdevice_context);
-
-	virtual void Render() const;
-
-	~OBJModel();
-};
 
 #endif
