+++ src/OBJModel.cpp
@@ -1,95 +1,5 @@
-//
-//  Model.cpp
-//
-//  CJ Gribel 2016, cjgribel@gmail.com
-//
-
-#include "Model.h"
-
-QuadModel::QuadModel(
-	ID3D11Device* dxdevice,
-	ID3D11DeviceContext* dxdevice_context)
-	: Model(dxdevice, dxdevice_context)
-{
-	// Vertex and index arrays
-	// Once their data is loaded to GPU buffers, they are not needed anymore
-	std::vector<Vertex> vertices;
-	std::vector<unsigned> indices;
-
-	// Populate the vertex array with 4 vertices
-	Vertex v0, v1, v2, v3;
-	v0.Pos = { -0.5, -0.5f, 0.0f };
-	v0.Normal = { 0, 0, 1 };
-	v0.TexCoord = { 0, 0 };
-	v1.Pos = { 0.5, -0.5f, 0.0f };
-	v1.Normal = { 0, 0, 1 };
-	v1.TexCoord = { 0, 1 };
-	v2.Pos = { 0.5, 0.5f, 0.0f };
-	v2.Normal = { 0, 0, 1 };
-	v2.TexCoord = { 1, 1 };
-	v3.Pos = { -0.5, 0.5f, 0.0f };
-	v3.Normal = { 0, 0, 1 };
-	v3.TexCoord = { 1, 0 };
-	vertices.push_back(v0);
-	vertices.push_back(v1);
-	vertices.push_back(v2);
-	vertices.push_back(v3);
-
-	// Populate the index array with two triangles
-	// Triangle #1
-	indices.push_back(0);
-	indices.push_back(1);
-	indices.push_back(3);
-	// Triangle #2
-	indices.push_back(1);
-	indices.push_back(2);
-	indices.push_back(3);
-
-	// Vertex array descriptor
-	D3D11_BUFFER_DESC vbufferDesc = { 0 };
-	vbufferDesc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
-	vbufferDesc.CPUAccessFlags = 0;
-	vbufferDesc.Usage = D3D11_USAGE_DEFAULT;
-	vbufferDesc.MiscFlags = 0;
-	vbufferDesc.ByteWidth = (UINT)(vertices.size()*sizeof(Vertex));
-	// Data resource
-	D3D11_SUBRESOURCE_DATA vdata;
-	vdata.pSysMem = &vertices[0];
-	// Create vertex buffer on device using descriptor & data
-	dxdevice->CreateBuffer(&vbufferDesc, &vdata, &vertex_buffer);
-	SETNAME(vertex_buffer, "VertexBuffer");
-    
-	//  Index array descriptor
-	D3D11_BUFFER_DESC ibufferDesc = { 0 };
-	ibufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
-	ibufferDesc.CPUAccessFlags = 0;
-	ibufferDesc.Usage = D3D11_USAGE_DEFAULT;
-	ibufferDesc.MiscFlags = 0;
-	ibufferDesc.ByteWidth = (UINT)(indices.size()*sizeof(unsigned));
-	// Data resource
-	D3D11_SUBRESOURCE_DATA idata;
-	idata.pSysMem = &indices[0];
-	// Create index buffer on device using descriptor & data
-	dxdevice->CreateBuffer(&ibufferDesc, &idata, &index_buffer);
-	SETNAME(index_buffer, "IndexBuffer");
-    
-	nbr_indices = (unsigned int)indices.size();
-}
-
-
-void QuadModel::Render() const
-{
-	// Bind our vertex buffer
-	const UINT32 stride = sizeof(Vertex); //  sizeof(float) * 8;
-	const UINT32 offset = 0;
-	dxdevice_context->IASetVertexBuffers(0, 1, &vertex_buffer, &stride, &offset);
-
-	// Bind our index buffer
-	dxdevice_context->IASetIndexBuffer(index_buffer, DXGI_FORMAT_R32_UINT, 0);
-
-	// Make the drawcall
-	dxdevice_context->DrawIndexed(nbr_indices, 0, 0);
-}
+#include "OBJModel.h"
+#include <future>
 
 OBJModel::OBJModel(
 	const std::string& objfile,
@@ -126,28 +36,28 @@ OBJModel::OBJModel(
 	vbufferDesc.CPUAccessFlags = 0;
 	vbufferDesc.Usage = D3D11_USAGE_DEFAULT;
 	vbufferDesc.MiscFlags = 0;
-	vbufferDesc.ByteWidth = (UINT)(mesh->vertices.size()*sizeof(Vertex));
+	vbufferDesc.ByteWidth = (UINT)(mesh->vertices.size() * sizeof(Vertex));
 	// Data resource
 	D3D11_SUBRESOURCE_DATA vdata;
 	vdata.pSysMem = &(mesh->vertices)[0];
 	// Create vertex buffer on device using descriptor & data
 	dxdevice->CreateBuffer(&vbufferDesc, &vdata, &vertex_buffer);
 	SETNAME(vertex_buffer, "VertexBuffer");
-    
+
 	// Index array descriptor
 	D3D11_BUFFER_DESC ibufferDesc = { 0 };
 	ibufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
 	ibufferDesc.CPUAccessFlags = 0;
 	ibufferDesc.Usage = D3D11_USAGE_DEFAULT;
 	ibufferDesc.MiscFlags = 0;
-	ibufferDesc.ByteWidth = (UINT)(indices.size()*sizeof(unsigned));
+	ibufferDesc.ByteWidth = (UINT)(indices.size() * sizeof(unsigned));
 	// Data resource
 	D3D11_SUBRESOURCE_DATA idata;
 	idata.pSysMem = &indices[0];
 	// Create index buffer on device using descriptor & data
 	dxdevice->CreateBuffer(&ibufferDesc, &idata, &index_buffer);
 	SETNAME(index_buffer, "IndexBuffer");
-    
+
 	// Copy materials from mesh
 	append_materials(mesh->materials);
 
@@ -163,9 +73,9 @@ OBJModel::OBJModel(
 
 			hr = LoadTextureFromFile(
 				dxdevice,
-				mtl.Kd_texture_filename.c_str(), 
+				mtl.Kd_texture_filename.c_str(),
 				&mtl.diffuse_texture);
-			std::cout << "\t" << mtl.Kd_texture_filename 
+			std::cout << "\t" << mtl.Kd_texture_filename
 				<< (SUCCEEDED(hr) ? " - OK" : "- FAILED") << std::endl;
 		}
 
@@ -176,7 +86,6 @@ OBJModel::OBJModel(
 
 	SAFE_DELETE(mesh);
 }
-
 
 void OBJModel::Render() const
 {
