+++ src/OBJModel.h
@@ -1,0 +1,43 @@
+/**
+ * @file OBJModel.h
+ * @brief Contains the OBJModel
+ * @author Carl Johan Gribel 2016, cjgribel@gmail.com
+*/
+
+#pragma once
+#include "Model.h"
+
+/**
+ * @brief Model representing a 3D object.
+ * @see OBJLoader
+*/
+class OBJModel : public Model
+{
+	// index ranges, representing drawcalls, within an index array
+	struct IndexRange
+	{
+		unsigned int start;
+		unsigned int size;
+		unsigned ofs;
+		int mtl_index;
+	};
+
+	std::vector<IndexRange> index_ranges;
+	std::vector<Material> materials;
+
+	void append_materials(const std::vector<Material>& mtl_vec)
+	{
+		materials.insert(materials.end(), mtl_vec.begin(), mtl_vec.end());
+	}
+
+public:
+
+	OBJModel(
+		const std::string& objfile,
+		ID3D11Device* dxdevice,
+		ID3D11DeviceContext* dxdevice_context);
+
+	virtual void Render() const;
+
+	~OBJModel();
+};
